path manager 구현(2) (2022-1-12)

+++ CStage_Play01.cpp
@@ -39,7 +39,7 @@ void CStage_Play01::Render(HDC _dc)
 {
 	CStage::Render(_dc);
 
-	TextOut(_dc, 10, 10, L"Here is Play01 Stage",wcslen(L"Here is Play01 Stage"));
+	TextOut(_dc, 10, 10, L"Here is Play01 Stage",(int)wcslen(L"Here is Play01 Stage"));
 }
 
 void CStage_Play01::Enter()
